Add missed casts for new test cases

+++ tests/output/stmts/do-while.c
@@ -26,7 +26,7 @@ static int cif_func5(void)
 #line 9 "/home/novikov/work/cif-gcc-11/tests/work/do-while.c.aux"
   return 0;
 #line 11 "/home/novikov/work/cif-gcc-11/tests/work/do-while.c.aux"
-  return ldv_func_res;
+  return ( int ) ldv_func_res;
 }
 #line 15 "/home/novikov/work/cif-gcc-11/tests/work/do-while.c.aux"
 static int cif_func6(void)
