+++ tests/output/stmts/switch.c
@@ -54,7 +54,7 @@ static int cif_func5(void)
 #line 9 "/home/novikov/work/cif-gcc-11/tests/work/switch.c.aux"
   return 0;
 #line 11 "/home/novikov/work/cif-gcc-11/tests/work/switch.c.aux"
-  return ldv_func_res;
+  return ( int ) ldv_func_res;
 }
 #line 15 "/home/novikov/work/cif-gcc-11/tests/work/switch.c.aux"
 static void cif_func3(void)
